shroten invitation code

+++ QuizWiz/Models/AccountViewModels.cs
@@ -143,7 +143,7 @@
         public string Email { get; set; }
 
         [Required]
-        public int ExamId { get; set; }
+        public long ExamId { get; set; }
 
         [Required]
         public string Code { get; set; }
